+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 All Notable changes to `flipboxfactory/craft-link` will be documented in this file
 
+## 1.1.0 - 2019-2-18
+### Added
+- Email link type
+- Enhanced readme/docs
+
+## 1.0.0 - 2019-2-17
+### Changed
+- Base namespacing from `flipbox\link` to `flipbox\craft\link`
+
 ## 1.0.0-beta.2 - 2017-11-10
 ### Added
 - Placeholder for external url inputs